requirements.txt
Django==2.2.4
django-appconf==1.0.3
django-bootstrap-form==3.4
django-dbbackup==3.2.0
django-user-accounts==2.1.0
pinax-documents==1.0.2
pinax-theme-bootstrap==8.0.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
